# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pet Hotel & Grooming System API",
    "description": "Contract-first OpenAPI for the Pet Hotel & Grooming System.\n\n**S1 status:** endpoints are stubs (HTTP 501) with complete request/response schemas — this document is the single source of truth for backend + frontend. Business logic, state machines and real payment/notification land in S2+.\n\nAuthoritative data model: `diagrams/puml/class_diagram_plantuml.puml`. Internal modules (payment, notification, audit) are not exposed as endpoints.",
    "version": "0.1.0-S1"
  },
  "paths": {
...
    "/api/v1/grooming/work-orders": {
      "get": {
        "tags": [
          "Grooming"
        ],
        "summary": "List work orders (groomer queue / by booking)",
        "operationId": "list_work_orders_api_v1_grooming_work_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkOrderOut"
                  },
                  "type": "array",
                  "title": "Response List Work Orders Api V1 Grooming Work Orders Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/grooming/work-orders/{work_order_id}": {
      "get": {
        "tags": [
          "Grooming"
        ],
        "summary": "Get work-order detail incl. photos (FR-04.4)",
        "operationId": "get_work_order_api_v1_grooming_work_orders__work_order_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/grooming/work-orders/{work_order_id}/start": {
      "post": {
        "tags": [
          "Grooming"
        ],
        "summary": "Start a work order: Pending → PreCheck (seq3)",
        "operationId": "start_work_order_api_v1_grooming_work_orders__work_order_id__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/grooming/work-orders/{work_order_id}/stage": {
      "post": {
        "tags": [
          "Grooming"
        ],
        "summary": "Advance stage: PreCheck/Bathing/Drying/Grooming (FR-04.5, seq3)",
        "operationId": "update_stage_api_v1_grooming_work_orders__work_order_id__stage_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/grooming/work-orders/{work_order_id}/photos": {
      "post": {
        "tags": [
          "Grooming"
        ],
        "summary": "Upload a work photo (FR-04.2)",
        "operationId": "upload_photo_api_v1_grooming_work_orders__work_order_id__photos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_photo_api_v1_grooming_work_orders__work_order_id__photos_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkPhotoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Grooming"
        ],
        "summary": "List work photos (FR-04.4)",
        "operationId": "list_photos_api_v1_grooming_work_orders__work_order_id__photos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkPhotoOut"
                  },
                  "title": "Response List Photos Api V1 Grooming Work Orders  Work Order Id  Photos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/grooming/work-orders/{work_order_id}/complete": {
      "post": {
        "tags": [
          "Grooming"
        ],
        "summary": "Complete a work order: Grooming → Completed (seq3)",
        "operationId": "complete_api_v1_grooming_work_orders__work_order_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/grooming/work-orders/{work_order_id}/emergency": {
      "post": {
        "tags": [
          "Grooming"
        ],
        "summary": "Trigger emergency: any stage → Aborted (FR-06.3, seq3)",
        "operationId": "trigger_emergency_api_v1_grooming_work_orders__work_order_id__emergency_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyTriggerIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_photo_api_v1_grooming_work_orders__work_order_id__photos_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_photo_api_v1_grooming_work_orders__work_order_id__photos_post"
      },
...
      "EmergencyEventOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "bookingId": {
            "type": "string",
            "format": "uuid",
            "title": "Bookingid"
          },
          "petId": {
            "type": "string",
            "format": "uuid",
            "title": "Petid"
          },
          "workOrderId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workorderid"
          },
          "reportedById": {
            "type": "string",
            "format": "uuid",
            "title": "Reportedbyid"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "occurredAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurredat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "bookingId",
          "petId",
          "reportedById",
          "description"
        ],
        "title": "EmergencyEventOut",
        "description": "Booking-layer event (SDD §8). workOrderId is set only for grooming-stage\nemergencies; NULL for front-desk/lodging ones."
      },
      "EmergencyTriggerIn": {
        "properties": {
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description",
            "examples": [
              "寵物突發抽搐，已送醫"
            ]
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "EmergencyTriggerIn",
        "description": "FR-06.3 — description is mandatory."
      },
...
      "GroomingStage": {
        "type": "string",
        "enum": [
          "PreCheck",
          "Bathing",
          "Drying",
          "Grooming"
        ],
        "title": "GroomingStage",
        "description": "The 4 advanceable stages (FR-04.5) — a subset of WorkStatus used as the\nrequest body for stage updates."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StageUpdateIn": {
        "properties": {
          "stage": {
            "$ref": "#/components/schemas/GroomingStage",
            "examples": [
              "Bathing"
            ]
          }
        },
        "type": "object",
        "required": [
          "stage"
        ],
        "title": "StageUpdateIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkOrderDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "bookingItemId": {
            "type": "string",
            "format": "uuid",
            "title": "Bookingitemid"
          },
          "assignedToId": {
            "type": "string",
            "format": "uuid",
            "title": "Assignedtoid"
          },
          "status": {
            "$ref": "#/components/schemas/WorkStatus"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/WorkPhotoOut"
            },
            "type": "array",
            "title": "Photos"
          },
          "emergencyEvent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmergencyEventOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "bookingItemId",
          "assignedToId",
          "status",
          "createdAt"
        ],
        "title": "WorkOrderDetailOut"
      },
      "WorkOrderOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "bookingItemId": {
            "type": "string",
            "format": "uuid",
            "title": "Bookingitemid"
          },
          "assignedToId": {
            "type": "string",
            "format": "uuid",
            "title": "Assignedtoid"
          },
          "status": {
            "$ref": "#/components/schemas/WorkStatus"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "bookingItemId",
          "assignedToId",
          "status",
          "createdAt"
        ],
        "title": "WorkOrderOut"
      },
      "WorkPhotoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workOrderId": {
            "type": "string",
            "format": "uuid",
            "title": "Workorderid"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "uploadedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploadedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workOrderId",
          "url"
        ],
        "title": "WorkPhotoOut"
      },
      "WorkStatus": {
        "type": "string",
        "enum": [
          "Pending",
          "PreCheck",
          "Bathing",
          "Drying",
          "Grooming",
          "Completed",
          "Aborted"
        ],
        "title": "WorkStatus"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Bearer JWT (enforced in S2)",
        "scheme": "bearer"
      }
    }
  },
...
}